+++ src/common/utils/common.cpp
@@ -15,12 +15,4 @@ const std::string& get_service_string(const role& service_role) {
         throw std::invalid_argument("invalid role");
 }
 
-uh::cluster::role get_service_role(const std::string& service_role_str) {
-    if (auto search = role_by_abbreviation.find(service_role_str);
-        search != role_by_abbreviation.end())
-        return search->second;
-
-    throw std::invalid_argument("invalid role");
-}
-
 } // namespace uh::cluster